+++ src/apple-pay/apple-pay.js
@@ -28,6 +28,10 @@ var wrapPromise = require("@braintree/wrap-promise");
  * @property {string} binData.payroll Possible values: 'Yes', 'No', 'Unknown'.
  * @property {string} binData.prepaid Possible values: 'Yes', 'No', 'Unknown'.
  * @property {string} binData.productId The product id.
+ * @property {string} binData.business Possible values: 'Yes', 'No', 'Unknown'.
+ * @property {string} binData.consumer Possible values: 'Yes', 'No', 'Unknown'.
+ * @property {string} binData.purchase Possible values: 'Yes', 'No', 'Unknown'.
+ * @property {string} binData.corporate Possible values: 'Yes', 'No', 'Unknown'.
  */
 
 /**