+++ src/client/request/graphql/generators/credit-card-for-fastlane-tokenization.js
@@ -35,6 +35,10 @@ function createMutation(config) {
     "        issuingBank " +
     "        countryOfIssuance " +
     "        productId " +
+    "        business " +
+    "        consumer " +
+    "        purchase " +
+    "        corporate " +
     "      } " +
     "        } " +
     "      } " +