+++ src/client/request/graphql/generators/credit-card-tokenization.js
@@ -31,6 +31,10 @@ function createMutation(config) {
     "        issuingBank " +
     "        countryOfIssuance " +
     "        productId " +
+    "        business " +
+    "        consumer " +
+    "        purchase " +
+    "        corporate " +
     "      } " +
     "    } ";