+++ src/google-payment/google-payment.js
@@ -36,6 +36,10 @@ var CREATE_PAYMENT_DATA_REQUEST_METHODS = {
  * @property {string} binData.payroll Possible values: 'Yes', 'No', 'Unknown'.
  * @property {string} binData.prepaid Possible values: 'Yes', 'No', 'Unknown'.
  * @property {string} binData.productId The product id.
+ * @property {string} binData.business Possible values: 'Yes', 'No', 'Unknown'.
+ * @property {string} binData.consumer Possible values: 'Yes', 'No', 'Unknown'.
+ * @property {string} binData.purchase Possible values: 'Yes', 'No', 'Unknown'.
+ * @property {string} binData.corporate Possible values: 'Yes', 'No', 'Unknown'.
  */
 
 /**