+++ test/google-payment/unit/google-payment.js
@@ -646,7 +646,7 @@ describe("GooglePayment", () => {
         cardInfo: {},
         paymentMethodToken: {
           token:
-            '{"androidPayCards":[{"type":"AndroidPayCard","nonce":"nonce-1234","description":"Android Pay","consumed":false,"details":{"cardType":"Visa","lastTwo":"24","lastFour":"7224","isNetworkTokenized":false,"bin":"411111"},"binData":{"prepaid":"No","healthcare":"Unknown","debit":"Unknown","durbinRegulated":"Unknown","commercial":"Unknown","payroll":"Unknown","issuingBank":"Unknown","countryOfIssuance":"","productId":"Unknown"}}]}',
+            '{"androidPayCards":[{"type":"AndroidPayCard","nonce":"nonce-1234","description":"Android Pay","consumed":false,"details":{"cardType":"Visa","lastTwo":"24","lastFour":"7224","isNetworkTokenized":false,"bin":"411111"},"binData":{"prepaid":"No","healthcare":"Unknown","debit":"Unknown","durbinRegulated":"Unknown","commercial":"Unknown","payroll":"Unknown","issuingBank":"Unknown","countryOfIssuance":"","productId":"Unknown","business":"Unknown","consumer":"Unknown","purchase":"Unknown","corporate":"Unknown"}}]}',
           tokenizationType: "PAYMENT_GATEWAY",
         },
       };
@@ -675,6 +675,10 @@ describe("GooglePayment", () => {
               issuingBank: "Unknown",
               countryOfIssuance: "",
               productId: "Unknown",
+              business: "Unknown",
+              consumer: "Unknown",
+              purchase: "Unknown",
+              corporate: "Unknown",
             },
           });
         });